Improve module structure
