+++ zkfold-base.cabal
@@ -120,7 +120,11 @@ library
       ZkFold.Base.Protocol.Plonkup.LookupConstraint
       ZkFold.Base.Protocol.Plonkup.PlonkConstraint
       ZkFold.Base.Protocol.Plonkup.Internal
+      ZkFold.Base.Protocol.Plonkup.Prover
       ZkFold.Base.Protocol.Plonkup.Relation
+      ZkFold.Base.Protocol.Plonkup.Setup
+      ZkFold.Base.Protocol.Plonkup.Utils
+      ZkFold.Base.Protocol.Plonkup.Verifier
       ZkFold.Base.Protocol.Protostar
       ZkFold.Base.Protocol.Protostar.Accumulator
       ZkFold.Base.Protocol.Protostar.AccumulatorScheme
